+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Since v0.10.0 this project adheres to [Semantic Versioning](http://semver.org/) 
 
 #### Added
 
+* Support for the maxlag parameter (with retries) in API requests ([#112])
 * Support for getting/setting user agent for API requests ([#107])
 * Added missing PHPDoc comments for properties, constants, and more ([#109])
 
@@ -129,4 +130,5 @@ Since v0.10.0 this project adheres to [Semantic Versioning](http://semver.org/) 
 [#108]: https://github.com/hamstar/Wikimate/pull/108
 [#109]: https://github.com/hamstar/Wikimate/pull/109
 [#111]: https://github.com/hamstar/Wikimate/pull/111
+[#112]: https://github.com/hamstar/Wikimate/pull/112
 [#114]: https://github.com/hamstar/Wikimate/pull/114